pyproject.toml
[tool.poetry]
name = "Barbara"
version = "0.1.0"
description = ""
authors = ["Chtholly2000 || https://github.com/Chtholly2000"]

[tool.poetry.dependencies]
python = "^3.8"
colorama = "^0.4.4"
"discord.py" = "1.7.3"
Flask = "^2.0.1"
pytz = "^2021.3"
DiscordUtils = "^1.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
